Edited files to fit new directory tree/structure
